+++ hook/hook.go
@@ -1,12 +1,17 @@
 package hook
 
 import (
+	"crypto/hmac"
+	"crypto/sha1"
+	"encoding/hex"
 	"encoding/json"
+	"fmt"
 	"io/ioutil"
 	"log"
+	"reflect"
 	"regexp"
-
-	"github.com/adnanh/webhook/helpers"
+	"strconv"
+	"strings"
 )
 
 // Constants used to specify the parameter source
@@ -15,6 +20,57 @@ const (
 	SourceQuery   string = "url"
 	SourcePayload string = "payload"
 )
+
+// CheckPayloadSignature calculates and verifies SHA1 signature of the given payload
+func CheckPayloadSignature(payload []byte, secret string, signature string) (string, bool) {
+	if strings.HasPrefix(signature, "sha1=") {
+		signature = signature[5:]
+	}
+
+	mac := hmac.New(sha1.New, []byte(secret))
+	mac.Write(payload)
+	expectedMAC := hex.EncodeToString(mac.Sum(nil))
+
+	return expectedMAC, hmac.Equal([]byte(signature), []byte(expectedMAC))
+}
+
+// ExtractParameter extracts value from interface{} based on the passed string
+func ExtractParameter(s string, params interface{}) (string, bool) {
+	if params == nil {
+		return "", false
+	}
+
+	if paramsValue := reflect.ValueOf(params); paramsValue.Kind() == reflect.Slice {
+		if paramsValueSliceLength := paramsValue.Len(); paramsValueSliceLength > 0 {
+
+			if p := strings.SplitN(s, ".", 2); len(p) > 1 {
+				index, err := strconv.ParseInt(p[0], 10, 64)
+
+				if err != nil {
+					return "", false
+				} else if paramsValueSliceLength <= int(index) {
+					return "", false
+				}
+
+				return ExtractParameter(p[1], params.([]interface{})[index])
+			}
+		}
+
+		return "", false
+	}
+
+	if p := strings.SplitN(s, ".", 2); len(p) > 1 {
+		if pValue, ok := params.(map[string]interface{})[p[0]]; ok {
+			return ExtractParameter(p[1], pValue)
+		}
+	} else {
+		if pValue, ok := params.(map[string]interface{})[p[0]]; ok {
+			return fmt.Sprintf("%v", pValue), true
+		}
+	}
+
+	return "", false
+}
 
 // Argument type specifies the parameter key name and the source it should
 // be extracted from
@@ -38,7 +94,7 @@ func (ha *Argument) Get(headers, query, payload *map[string]interface{}) (string
 	}
 
 	if source != nil {
-		return helpers.ExtractParameter(ha.Name, *source)
+		return ExtractParameter(ha.Name, *source)
 	}
 
 	return "", false
@@ -201,7 +257,7 @@ func (r MatchRule) Evaluate(headers, query, payload *map[string]interface{}, bod
 			}
 			return ok
 		case MatchHashSHA1:
-			expected, ok := helpers.CheckPayloadSignature(*body, r.Secret, arg)
+			expected, ok := CheckPayloadSignature(*body, r.Secret, arg)
 			if !ok {
 				log.Printf("payload signature mismatch, expected %s got %s", expected, arg)
 			}
