+++ hook/hook_test.go
@@ -1,4 +1,4 @@
-package helpers
+package hook
 
 import "testing"
